scraping/requirements.txt
# pandas
# requests
# bs4
# retry
# PyYaml
# duckdb
# gspread
# google
# gspread_dataframe

beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
decorator==5.2.1
duckdb==1.4.3
google==3.0.0
google-auth==2.47.0
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-dataframe==4.0.0
idna==3.11
numpy==2.4.1
oauthlib==3.3.1
pandas==2.3.3
py==1.11.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
retry==0.9.2
rsa==4.9.1
six==1.17.0
soupsieve==2.8.1
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3